Refactor detection for DHCP configuration call
When the network interface is IPv6 only, no need for a DHCP call.
This is only usefull for network interface with IPv4 configuration.

If the network interface is attached to the VM after the boot/creation,
its information are not present in the JSON config file. We start DHCP
client for all post-boot attached network interface.

We use different return code for each cases in the JSON parser.

Signed-off-by: aegiap <aegiap@gandi.net>