Install autofs setup and avoid udev duplicate
For Debian stretch we use a setup based on autofs
to handle the virtual disk mount and umount. To avoid
conflict with our udev rules, we disable them in the
udev config.

We need to plan a proper migration of udev rules file
from /etc/udev/rules.d to /lib/udev/rules.d and for now
we should avoid duplicate files with the same set of rules.

Also remove the call to the Mandriva specific script.

Signed-off-by: aegiap <aegiap@gandi.net>