manage_iface: Add support for sandboxed systemd-udev
systemd-udev limits the actions a udev script can do by putting
it in a sandbox. This removes the possibility for the udev hook
to call dhclient.

In order to comply with systemd-udev requirements a dhclient generic
service is created that can be started by the udev hook.
The service is bound to the device so that the service is automatically
stopped if the interface is removed.

One notable change from the previous in-script mechanism is the use of a
long-lived dhclient instead of a one-shot instance.

Signed-off-by: Nicolas Belouin <nicolas.belouin@gandi.net>